+++ backend/tests/baserow/contrib/database/api/fields/test_field_views.py
@@ -168,7 +168,7 @@ def test_update_field(api_client, data_fixture):
     user_2, token_2 = data_fixture.create_user_and_token()
     table = data_fixture.create_database_table(user=user)
     table_2 = data_fixture.create_database_table(user=user_2)
-    text = data_fixture.create_text_field(table=table)
+    text = data_fixture.create_text_field(table=table, primary=True)
     text_2 = data_fixture.create_text_field(table=table_2)
 
     url = reverse('api:database:fields:item', kwargs={'field_id': text_2.id})
@@ -191,6 +191,22 @@ def test_update_field(api_client, data_fixture):
     )
     assert response.status_code == HTTP_404_NOT_FOUND
     assert response.json()['error'] == 'ERROR_FIELD_DOES_NOT_EXIST'
+
+    # The primary field is not compatible with a link row field so that should result
+    # in an error.
+    url = reverse('api:database:fields:item', kwargs={'field_id': text.id})
+    response = api_client.patch(
+        url,
+        {'type': 'link_row'},
+        format='json',
+        HTTP_AUTHORIZATION=f'JWT {token}'
+    )
+    assert response.status_code == HTTP_400_BAD_REQUEST
+    assert response.json()['error'] == 'ERROR_INCOMPATIBLE_PRIMARY_FIELD_TYPE'
+    assert (
+        response.json()['detail'] ==
+        'The field type link_row is not compatible with the primary field.'
+    )
 
     url = reverse('api:database:fields:item', kwargs={'field_id': text.id})
     response = api_client.patch(