+++ changelog.md
@@ -11,6 +11,8 @@
 * Set un-secure lax cookie when public web frontend url isn't over a secure connection.
 * Fixed bug where the sort choose field item didn't have a hover effect.
 * Implemented a file field and user files upload.
+* Made it impossible for the `link_row` field to be a primary field because that can
+  cause the primary field to be deleted.
 
 ## Released (2020-11-02)